Extract HarnessSessionFacade from WorkflowService
Moves the harness runtime/sensor/task/policy/workflow-state service
ownership plus the session-lifecycle methods (getHarnessStatus,
checkpointHarnessSession, recordHarnessArtifact, defineHarnessTask,
runHarnessSensors, ensureHarnessSession, registerDefaultSensors,
shell sensor execution) into a dedicated facade. WorkflowService
delegates, dropping from 1132 to ~785 lines without behavior changes.
