+++ src/services/workflow/workflowService.ts
@@ -6,11 +6,9 @@
 
 import * as path from 'path';
 import * as fs from 'fs-extra';
-import { exec as execCallback } from 'child_process';
-import { promisify } from 'util';
 import { CollaborationSession, CollaborationManager } from '../../workflow/collaboration';
 import { GateCheckResult, ExecutionEvidence } from '../../workflow/gates';
-import { PHASE_NAMES_PT } from '../../workflow/phases';
+import { PHASE_NAMES_EN, PHASE_NAMES_PT } from '../../workflow/phases';
 import { createPlanLinker, type LinkedPlan } from '../../workflow/plans';
 import { ROLE_DISPLAY_NAMES } from '../../workflow/roles';
 import { getScaleName, getScaleFromName } from '../../workflow/scaling';
@@ -27,46 +25,17 @@ import {
 } from '../../workflow/types';
 import type { CollaborationSynthesis } from '../../workflow/types';
 import type { PlanLinker } from '../../workflow/plans';
-import { PHASE_NAMES_EN } from '../../workflow/phases';
 import { FileCollaborationStore } from './fileCollaborationStore';
 import { DerivedPlanTaskContractBuilder } from './derivedPlanTaskContractBuilder';
 import {
-  HarnessRuntimeStateService,
-  HarnessSensorCatalogService,
-  HarnessSensorsService,
-  HarnessTaskContractsService,
-  HarnessPolicyService,
-  HarnessWorkflowStateService,
   type HarnessArtifactKind,
   type HarnessTaskContract,
-  type HarnessHandoffContract,
-  type HarnessSensorRun,
-  type HarnessSessionRecord,
   type HarnessSensorDefinition,
   type WorkflowHarnessBinding,
 } from '../harness';
+import { HarnessSessionFacade, type WorkflowHarnessStatus } from './harnessSessionFacade';
 
-const exec = promisify(execCallback);
-
-export interface WorkflowHarnessStatus {
-  binding: WorkflowHarnessBinding;
-  session: HarnessSessionRecord;
-  availableSensors: Array<Pick<HarnessSensorDefinition, 'id' | 'name' | 'description'>>;
-  sensorRuns: HarnessSensorRun[];
-  taskContracts: HarnessTaskContract[];
-  handoffs: HarnessHandoffContract[];
-  policyRules: number;
-  completionCheck: {
-    blocked: boolean;
-    reasons: string[];
-    taskCompletion: {
-      canComplete: boolean;
-      missingSensors: string[];
-      missingArtifacts: string[];
-      blockingFindings: string[];
-    } | null;
-  };
-}
+export type { WorkflowHarnessStatus };
 
 export class HarnessWorkflowBlockedError extends Error {
   constructor(
@@ -124,12 +93,7 @@ export class WorkflowService {
   private contextPath: string;
   private orchestrator: PrevcOrchestrator;
   private collaborationManager: CollaborationManager;
-  private runtimeStateService: HarnessRuntimeStateService;
-  private sensorCatalogService: HarnessSensorCatalogService;
-  private sensorsService: HarnessSensorsService;
-  private taskContractsService: HarnessTaskContractsService;
-  private policyService: HarnessPolicyService;
-  private workflowStateService: HarnessWorkflowStateService;
+  private harness: HarnessSessionFacade;
   private derivedTaskBuilder: DerivedPlanTaskContractBuilder;
   private deps: WorkflowServiceDependencies;
 
@@ -144,26 +108,24 @@ export class WorkflowService {
     this.contextPath = path.basename(resolvedPath) === '.context'
       ? resolvedPath
       : path.join(resolvedPath, '.context');
-    this.runtimeStateService = new HarnessRuntimeStateService({ repoPath: this.repoPath });
-    this.sensorCatalogService = new HarnessSensorCatalogService({
+    this.harness = new HarnessSessionFacade({
       repoPath: this.repoPath,
       contextPath: this.contextPath,
     });
-    this.sensorsService = new HarnessSensorsService({ stateService: this.runtimeStateService });
-    this.taskContractsService = new HarnessTaskContractsService({
-      repoPath: this.repoPath,
-      stateService: this.runtimeStateService,
-    });
-    this.policyService = new HarnessPolicyService({ repoPath: this.repoPath });
-    this.workflowStateService = new HarnessWorkflowStateService({ contextPath: this.contextPath });
-    this.orchestrator = new PrevcOrchestrator(this.contextPath, this.workflowStateService);
+    this.orchestrator = new PrevcOrchestrator(this.contextPath, this.harness.workflowStateService);
     this.collaborationManager = new CollaborationManager(
       new FileCollaborationStore(this.contextPath)
     );
     this.derivedTaskBuilder = new DerivedPlanTaskContractBuilder();
     this.deps = deps;
-    this.registerDefaultSensors();
   }
+
+  // Narrow accessors used by advance/handoff/plan-sync logic below.
+  // Fresh `get` each call keeps the facade as the single owner.
+  private get runtimeStateService() { return this.harness.runtimeStateService; }
+  private get sensorsService() { return this.harness.sensorsService; }
+  private get taskContractsService() { return this.harness.taskContractsService; }
+  private get policyService() { return this.harness.policyService; }
 
   /**
    * Create a WorkflowService with the given repository path
@@ -237,7 +199,7 @@ export class WorkflowService {
       options.archivePrevious
     );
 
-    await this.ensureHarnessSession(options.name, options.description);
+    await this.harness.ensureHarnessSession(options.name, options.description);
 
     this.deps.ui?.displaySuccess(
       `Workflow PREVC initialized: ${options.name} (Scale: ${getScaleName(scale)})`
@@ -260,31 +222,14 @@ export class WorkflowService {
     return this.orchestrator.getSummary();
   }
 
-  /**
-   * Get the English display name for a PREVC phase.
-   *
-   * Facade method so MCP/CLI handlers do not need to import phase
-   * constants directly from the workflow barrel.
-   */
   getPhaseDisplayName(phase: PrevcPhase): string {
     return PHASE_NAMES_EN[phase];
   }
 
-  /**
-   * Get the English display name for a PREVC role.
-   *
-   * Facade method for MCP/CLI handlers.
-   */
   getRoleDisplayName(role: PrevcRole): string {
     return ROLE_DISPLAY_NAMES[role];
   }
 
-  /**
-   * Return a PlanLinker bound to this workflow's repository.
-   *
-   * Facade method so MCP/CLI handlers never import `createPlanLinker`
-   * directly from the workflow submodules.
-   */
   getPlanLinkerForWorkflow(): PlanLinker {
     return createPlanLinker(this.repoPath);
   }
@@ -380,10 +325,6 @@ export class WorkflowService {
       : null;
     const plannedNextPhase = await this.orchestrator.getNextPhase();
 
-    // Evaluate execution evidence for the current active task BEFORE advancing.
-    // The execution_evidence gate (in gateChecker) will block transitions that
-    // leave required sensors/artifacts unsatisfied, even when autonomous_mode
-    // is on — autonomous only suppresses policy gates (plan/approval).
     const executionEvidence = await this.buildExecutionEvidence(
       activeTaskBeforeAdvance,
       binding.sessionId
@@ -449,16 +390,10 @@ export class WorkflowService {
     return nextPhase;
   }
 
-  /**
-   * Check workflow gates for the current phase transition
-   */
   async checkGates(): Promise<GateCheckResult> {
     return this.orchestrator.checkGates();
   }
 
-  /**
-   * Set workflow settings
-   */
   async setSettings(settings: Partial<WorkflowSettings>): Promise<WorkflowSettings> {
     const isHighRisk =
       typeof settings.autonomous_mode === 'boolean' ||
@@ -472,24 +407,14 @@ export class WorkflowService {
     return this.orchestrator.setSettings(settings);
   }
 
-  /**
-   * Get workflow settings
-   */
   async getSettings(): Promise<WorkflowSettings> {
     return this.orchestrator.getSettings();
   }
 
   listAvailableSensors(): Array<Pick<HarnessSensorDefinition, 'id' | 'name' | 'description'>> {
-    return this.sensorsService.listSensors().map((sensor) => ({
-      id: sensor.id,
-      name: sensor.name,
-      description: sensor.description,
-    }));
+    return this.harness.listAvailableSensors();
   }
 
-  /**
-   * Enable or disable autonomous mode
-   */
   async setAutonomousMode(enabled: boolean): Promise<WorkflowSettings> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -500,9 +425,6 @@ export class WorkflowService {
     return this.orchestrator.setSettings({ autonomous_mode: enabled });
   }
 
-  /**
-   * Mark that a plan has been created/linked
-   */
   async markPlanCreated(planSlug: string): Promise<void> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -639,9 +561,6 @@ export class WorkflowService {
     });
   }
 
-  /**
-   * Approve the plan
-   */
   async approvePlan(approver: PrevcRole | string, notes?: string): Promise<PlanApproval> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -653,16 +572,10 @@ export class WorkflowService {
     return this.orchestrator.approvePlan(approver, notes);
   }
 
-  /**
-   * Get approval status
-   */
   async getApproval(): Promise<PlanApproval | undefined> {
     return this.orchestrator.getApproval();
   }
 
-  /**
-   * Perform a handoff between agents
-   */
   async handoff(
     from: string,
     to: string,
@@ -712,23 +625,14 @@ export class WorkflowService {
     );
   }
 
-  /**
-   * Get orchestration guidance for a phase
-   */
   async getPhaseOrchestration(phase: PrevcPhase): Promise<PhaseOrchestration> {
     return this.orchestrator.getPhaseOrchestration(phase);
   }
 
-  /**
-   * Get next agent suggestion after a handoff
-   */
   getNextAgentSuggestion(currentAgent: string): { agent: string; reason: string } | null {
     return this.orchestrator.getNextAgentSuggestion(currentAgent);
   }
 
-  /**
-   * Start a collaboration session
-   */
   async startCollaboration(
     topic: string,
     participants?: PrevcRole[]
@@ -745,9 +649,6 @@ export class WorkflowService {
     return session;
   }
 
-  /**
-   * Add a contribution to a collaboration session
-   */
   contributeToCollaboration(
     sessionId: string,
     role: PrevcRole,
@@ -756,30 +657,18 @@ export class WorkflowService {
     this.collaborationManager.contribute(sessionId, role, message);
   }
 
-  /**
-   * End a collaboration session and get synthesis
-   */
   async endCollaboration(sessionId: string): Promise<CollaborationSynthesis | null> {
     return this.collaborationManager.endSession(sessionId);
   }
 
-  /**
-   * Get recommended next actions
-   */
   async getRecommendedActions(): Promise<string[]> {
     return this.orchestrator.getRecommendedActions();
   }
 
-  /**
-   * Check if workflow is complete
-   */
   async isComplete(): Promise<boolean> {
     return this.orchestrator.isComplete();
   }
 
-  /**
-   * Update the current task
-   */
   async updateTask(task: string): Promise<void> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -790,9 +679,6 @@ export class WorkflowService {
     await this.orchestrator.updateCurrentTask(task);
   }
 
-  /**
-   * Start a role in the current phase
-   */
   async startRole(role: PrevcRole): Promise<void> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -806,9 +692,6 @@ export class WorkflowService {
     );
   }
 
-  /**
-   * Complete a role's work
-   */
   async completeRole(role: PrevcRole, outputs: string[]): Promise<void> {
     await this.policyService.authorize({
       tool: 'workflow',
@@ -830,56 +713,8 @@ export class WorkflowService {
     if (!(await this.hasWorkflow())) {
       return null;
     }
-
     const summary = await this.getSummary();
-    const binding = await this.ensureHarnessSession(summary.name);
-
-    const session = await this.runtimeStateService.getSession(binding.sessionId);
-    const allSensorRuns = await this.sensorsService.getSessionSensorRuns(binding.sessionId);
-    const policyRules = await this.policyService.listRules();
-    const latestRuns = new Map<string, HarnessSensorRun>();
-    for (const run of allSensorRuns) {
-      const current = latestRuns.get(run.sensorId);
-      if (!current || current.createdAt < run.createdAt) {
-        latestRuns.set(run.sensorId, run);
-      }
-    }
-    const sensorRuns = Array.from(latestRuns.values()).sort((a, b) => a.sensorId.localeCompare(b.sensorId));
-    const taskContracts = (await this.taskContractsService.listTaskContracts())
-      .filter((contract) => contract.sessionId === binding.sessionId);
-    const handoffs = (await this.taskContractsService.listHandoffContracts())
-      .filter((handoff) => handoff.sessionId === binding.sessionId);
-    const activeTask = binding.activeTaskId
-      ? taskContracts.find((contract) => contract.id === binding.activeTaskId) ?? null
-      : null;
-    const taskCompletion = activeTask
-      ? await this.taskContractsService.evaluateTaskCompletion(activeTask.id, binding.sessionId)
-      : null;
-    const backpressure = this.sensorsService.evaluateBackpressure(sensorRuns, { requireEvidence: true });
-    const reasons = [
-      ...backpressure.reasons,
-      ...(taskCompletion?.blockingFindings || []),
-    ];
-
-    return {
-      binding,
-      session,
-      availableSensors: this.listAvailableSensors(),
-      sensorRuns,
-      taskContracts,
-      handoffs,
-      policyRules: policyRules.length,
-      completionCheck: {
-        blocked: reasons.length > 0,
-        reasons,
-        taskCompletion: taskCompletion ? {
-          canComplete: taskCompletion.canComplete,
-          missingSensors: taskCompletion.missingSensors,
-          missingArtifacts: taskCompletion.missingArtifacts,
-          blockingFindings: taskCompletion.blockingFindings,
-        } : null,
-      },
-    };
+    return this.harness.getHarnessStatus(summary.name);
   }
 
   async checkpointHarnessSession(
@@ -887,27 +722,9 @@ export class WorkflowService {
     data?: unknown,
     artifactIds?: string[],
     pause?: boolean
-  ): Promise<{ binding: WorkflowHarnessBinding; session: HarnessSessionRecord }> {
+  ) {
     const binding = await this.requireHarnessBinding();
-    await this.policyService.authorize({
-      tool: 'workflow',
-      action: 'checkpoint',
-      risk: pause ? 'high' : 'medium',
-      metadata: {
-        note,
-        pause: Boolean(pause),
-        artifactCount: artifactIds?.length ?? 0,
-      },
-    });
-    const session = await this.runtimeStateService.checkpointSession(binding.sessionId, {
-      note,
-      data,
-      artifactIds,
-      pause,
-    });
-    binding.updatedAt = session.updatedAt;
-    await this.saveHarnessBinding(binding);
-    return { binding, session };
+    return this.harness.checkpointHarnessSession(binding, { note, data, artifactIds, pause });
   }
 
   async recordHarnessArtifact(input: {
@@ -918,17 +735,7 @@ export class WorkflowService {
     metadata?: Record<string, unknown>;
   }) {
     const binding = await this.requireHarnessBinding();
-    await this.policyService.authorize({
-      tool: 'workflow',
-      action: 'recordArtifact',
-      paths: input.path ? [input.path] : [input.name],
-      risk: input.path?.includes('secret') ? 'critical' : 'medium',
-      metadata: input.metadata,
-    });
-    const artifact = await this.runtimeStateService.addArtifact(binding.sessionId, input);
-    binding.updatedAt = artifact.createdAt;
-    await this.saveHarnessBinding(binding);
-    return artifact;
+    return this.harness.recordHarnessArtifact(binding, input);
   }
 
   async defineHarnessTask(input: {
@@ -943,87 +750,17 @@ export class WorkflowService {
     metadata?: Record<string, unknown>;
   }): Promise<HarnessTaskContract> {
     const binding = await this.requireHarnessBinding();
-    await this.policyService.authorize({
-      tool: 'workflow',
-      action: 'defineTask',
-      risk: input.requiredSensors?.includes('deploy') ? 'high' : 'medium',
-      metadata: {
-        ...input.metadata,
-        title: input.title,
-      },
-    });
-    const contract = await this.taskContractsService.createTaskContract({
-      ...input,
-      sessionId: binding.sessionId,
-      status: 'ready',
-    });
-
-    binding.activeTaskId = contract.id;
-    binding.updatedAt = contract.updatedAt;
-    await this.saveHarnessBinding(binding);
-    return contract;
+    return this.harness.defineHarnessTask(binding, input);
   }
 
   async runHarnessSensors(sensorIds: string[], metadata?: Record<string, unknown>) {
     const binding = await this.requireHarnessBinding();
-    const runs: HarnessSensorRun[] = [];
-
-    for (const sensorId of sensorIds) {
-      await this.policyService.authorize({
-        tool: 'workflow',
-        action: 'runSensors',
-        risk: sensorId === 'deploy' ? 'high' : 'medium',
-        metadata,
-      });
-      runs.push(await this.sensorsService.runSensor(sensorId, {
-        sessionId: binding.sessionId,
-        contractId: binding.activeTaskId,
-        metadata,
-      }));
-    }
-
-    return {
-      runs,
-      backpressure: this.sensorsService.evaluateBackpressure(runs, { requireEvidence: true }),
-    };
-  }
-
-  private async ensureHarnessSession(
-    workflowName: string,
-    description?: string
-  ): Promise<WorkflowHarnessBinding> {
-    const existing = await this.loadHarnessBinding();
-    if (existing && existing.workflowName === workflowName) {
-      try {
-        await this.runtimeStateService.getSession(existing.sessionId);
-        return existing;
-      } catch {
-        // Session was removed or became unreadable. Recreate below.
-      }
-    }
-
-    const session = await this.runtimeStateService.createSession({
-      name: workflowName,
-      metadata: {
-        workflow: true,
-        description,
-      },
-    });
-
-    const binding: WorkflowHarnessBinding = {
-      workflowName,
-      sessionId: session.id,
-      createdAt: new Date().toISOString(),
-      updatedAt: session.updatedAt,
-    };
-
-    await this.saveHarnessBinding(binding);
-    return binding;
+    return this.harness.runHarnessSensors(binding, sensorIds, metadata);
   }
 
   private async requireHarnessBinding(): Promise<WorkflowHarnessBinding> {
     const summary = await this.getSummary();
-    return this.ensureHarnessSession(summary.name);
+    return this.harness.ensureHarnessSession(summary.name);
   }
 
   private async resolveLinkedPlan(planSlug?: string): Promise<LinkedPlan | null> {
@@ -1042,91 +779,7 @@ export class WorkflowService {
     return linker.getLinkedPlan(resolvedPlanSlug);
   }
 
-  private async loadHarnessBinding(): Promise<WorkflowHarnessBinding | null> {
-    return this.workflowStateService.getBinding();
-  }
-
   private async saveHarnessBinding(binding: WorkflowHarnessBinding): Promise<void> {
-    await this.workflowStateService.saveBinding(binding);
-  }
-
-  private registerDefaultSensors(): void {
-    const definitions = this.sensorCatalogService.resolveEffectiveSensorsSync();
-
-    for (const definition of definitions) {
-      if (this.sensorsService.getSensor(definition.id)) {
-        continue;
-      }
-
-      this.sensorsService.registerSensor({
-        id: definition.id,
-        name: definition.name,
-        severity: definition.severity,
-        blocking: definition.severity === 'critical',
-        execute: async () => {
-          if (definition.script) {
-            const hasScript = await this.hasPackageScript(definition.script);
-            if (!hasScript) {
-              return {
-                status: definition.severity === 'warning' ? 'skipped' : 'blocked',
-                summary: `Script not available: ${definition.script}`,
-                evidence: [`Missing package.json script: ${definition.script}`],
-              };
-            }
-          }
-
-          return this.executeShellSensor(definition.command, definition.name);
-        },
-      });
-    }
-  }
-
-  private async hasPackageScript(scriptName: string): Promise<boolean> {
-    const packageJsonPath = path.join(this.repoPath, 'package.json');
-    if (!(await fs.pathExists(packageJsonPath))) {
-      return false;
-    }
-
-    const packageJson = await fs.readJson(packageJsonPath) as { scripts?: Record<string, string> };
-    return Boolean(packageJson.scripts?.[scriptName]);
-  }
-
-  private async executeShellSensor(command: string, name: string) {
-    try {
-      const { stdout, stderr } = await exec(command, {
-        cwd: this.repoPath,
-        timeout: 300000,
-      });
-
-      return {
-        status: 'passed' as const,
-        summary: `${name} passed`,
-        evidence: [this.trimEvidence(stdout), this.trimEvidence(stderr)].filter(Boolean) as string[],
-        output: {
-          command,
-        },
-      };
-    } catch (error: any) {
-      return {
-        status: 'failed' as const,
-        summary: `${name} failed`,
-        evidence: [
-          this.trimEvidence(error?.stdout),
-          this.trimEvidence(error?.stderr || error?.message),
-        ].filter(Boolean) as string[],
-        details: {
-          command,
-          exitCode: typeof error?.code === 'number' ? error.code : undefined,
-        },
-      };
-    }
-  }
-
-  private trimEvidence(value?: string, maxLength: number = 2000): string | null {
-    if (!value) {
-      return null;
-    }
-
-    return value.length > maxLength ? `${value.slice(0, maxLength)}\n...[truncated]` : value;
+    await this.harness.saveHarnessBinding(binding);
   }
 }
